requirements.txt
openai
Pillow
requests
packaging
customtkinter
